refactor Graphics code
